requirements.txt
h3
google-cloud-bigquery==3.4.1
apache-airflow~=2.5.0
apache-airflow-providers-slack==7.1.1
apache-airflow-providers-google
protobuf==4.21.12
google-cloud-storage==2.7.0
pytz==2022.7.1
dependencies~=7.7.0
requests==2.28.2
pandas~=1.5.2
